Components were not properly tracking which view was active
Just had to un-bind a setView() from a the Notes component, where it wasn't supposed to be used as well as remove the reactive component that was being passed into NotesHome.  The issue was causing the store to not actually update for some reason which led to some views not being able to render when the nav button was clicked. Also, edited the CSS